Makefile for all stages Close #15 Began with dreaded cv2 not found. Having earned agony earlier in this project, I knew that the problem was most likely with running setupvars and init OpenVINO env.
Log in with a shell (as root, which prompted the run_root addition) and running
'source /opt/intel/openvino/bin/setupvars.sh' allowed me to then
run the classification sample. Actually, I had to unearth some dependencies first,
such as libsm6, libgtk... all the ones now included in the first RUN in Dockerfile.

The 'source' command in bash essentially makes variables and things set when the script
is executed available to later processes. Simply calling setupvars.sh didn't solve the
problem because then nothing was changed when I ran inference. However, source is not
included in sh, which is Docker default. Fortunately, '.' served enough of the same
purpose.

I opted to switch to . instead of using SHELL in the Dockerfile to switch the default.

I also added a '-l' to the CMD. This is to make sure that .bashrc gets sourced when
going into an interactive shell.
fatal: Invalid path '/opt/intel/openvino_2019.2.242/bin/setupvars.sh allowed me to then
run the classification sample. Actually, I had to unearth some dependencies first,
such as libsm6, libgtk... all the ones now included in the first RUN in Dockerfile.

The source command in bash essentially makes variables and things set when the script
is executed available to later processes. Simply calling setupvars.sh didn't solve the
problem because then nothing was changed when I ran inference. However, source is not
included in sh, which is Docker default. Fortunately, . served enough of the same
purpose.

I opted to switch to . instead of using SHELL in the Dockerfile to switch the default.

I also added a -l to the CMD. This is to make sure that .bashrc gets sourced when
going into an interactive shell.